test: split retrieval snapshot contracts
